refactor: `FossilExecutable` in separate file
